Add virtual notches for trackpads (via haptic feedback)
For Smash Bros., it's fairly important to know exactly what angle the joystick is pointing. On physical devices, one can use notches (such as the GameCube controller) to determine the angle of the joystick. This commit adds "virtual notches" by triggering haptic feedback when the joystick rotates by a certain amount. (Unlike physical notches, the virtual notches trigger even when the joystick is not pressed to the maximum amount, against the joystick shell.)

Note that the Steam Controller hardware has the trackpad sensors offset by about 15 degrees. This commit doesn't address that, so the notches will appear to be 15 degrees off.