Fix bug with last XY positioning
Before this commit, when the user completely lifts their finger, we retain their last XY position. Then, if the user taps a dead spot, we reuse that position. This is wrong, such as if the user last had their finger on the bottom of the trackpad and then they tap the top of the trackpad. This commit fixes it so that the last XY position is reset when the user lifts their finger.