Fix bug with inner ring size calculation
Before this commit, we used the normal norm-2 distance function to calculate the finger position's distance from the center, However, for whatever reason, the hardware actually produces X values in the range of 1-1200 but Y values in the range of 0-700. This means that we were effectively checking the finger position with respect to an ellipse, not a circle. This commit fixes that by normalizing the X and Y position with respect to the radius.